+++ src/sdsstools/__init__.py
@@ -22,6 +22,7 @@ from ._vendor import color_print, toml, yanny
 from .configuration import *
 from .logger import *
 from .metadata import *
+from .retrier import *
 from .time import get_sjd
 from .utils import *